issue #613 broadcast model
