+++ openeo-geotrellis/src/main/scala/org/openeo/geotrellis/OpenEOProcesses.scala
@@ -22,6 +22,7 @@ import geotrellis.spark.{MultibandTileLayerRDD, _}
 import geotrellis.util._
 import geotrellis.vector.Extent.toPolygon
 import geotrellis.vector._
+import org.apache.commons.io.FileUtils
 import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.rdd._
 import org.apache.spark.{Partitioner, SparkContext}
@@ -34,6 +35,7 @@ import org.openeo.geotrelliscommon.{ByTileSpacetimePartitioner, ByTileSpatialPar
 import org.slf4j.LoggerFactory
 
 import java.io.File
+import java.net.URL
 import java.nio.charset.StandardCharsets
 import java.nio.file.{Files, Paths}
 import java.time.format.DateTimeFormatter
@@ -1519,8 +1521,19 @@ class OpenEOProcesses extends Serializable {
 
   def predictONNXGeneric[K: SpatialComponent: ClassTag, M: Component[*, Bounds[K]]](datacube: MultibandTileLayerRDD[K], model:String): MultibandTileLayerRDD[K] = {
     logger.info("ONNX: start predictOnnx")
+    val modelPath = Paths.get(model)
+    val (modelFile, isTemp) = if (Files.exists(modelPath)) {
+      (modelPath,false)
+    } else {
+      val tempFileName = Files.createTempFile(null, ".onnx")
+      FileUtils.copyURLToFile(new URL(model), tempFileName.toFile)
+      (tempFileName,true)
+    }
+    val env = OrtEnvironment.getEnvironment()
+    val session = env.createSession(modelFile.toString, new OrtSession.SessionOptions())
+    val broadcastSession = sc.broadcast(session)
     ContextRDD(
-      datacube.mapValues(x => onnx.predictOnnx(x,model)),
+      datacube.mapValues(x => onnx.predictOnnx(x,broadcastSession.value)),
       datacube.metadata
     )
   }
