+++ openeo-geotrellis/src/main/scala/org/openeo/geotrellis/onnx/package.scala
@@ -118,18 +118,9 @@ package object onnx {
     inputArray
   }
 
-  def predictOnnx(tile: MultibandTile, model: String): MultibandTile = {
-    val modelPath = Paths.get(model)
-    val (modelFile, isTemp) = if (Files.exists(modelPath)) {
-      (modelPath,false)
-    } else {
-      val tempFileName = Files.createTempFile(null, ".onnx")
-      FileUtils.copyURLToFile(new URL(model), tempFileName.toFile)
-      (tempFileName,true)
-    }
+  def predictOnnx(tile: MultibandTile, session: OrtSession): MultibandTile = {
     val bandCount = tile.bandCount
     val env = OrtEnvironment.getEnvironment()
-    val session = env.createSession(modelFile.toString, new OrtSession.SessionOptions())
     val inputNames = session.getInputNames
     val outputNames = session.getOutputNames
 
@@ -170,7 +161,6 @@ package object onnx {
     val onnxResults = session.run(inputs)
     val resultValue = onnxResults.get(0).getValue.asInstanceOf[Array[_]]
     val flattenedResult = flattenNestedArray(resultValue, outputShape, outputType)
-    if (isTemp) Files.delete(modelFile)
     flattenedResult
   }
